Implement image cache.
Replaces two hardcoded Evas images o1 and o2 with an array,
maintained completely in the working thread.

This reduces the amount of confusion/code duplication
and lays foundations for displaying more than one page
on the screen (in the future).

Note that the cache replacement procedure is dumb,
and the 'page request queue' is not really a queue,
but the synchronization/threading seems sound.